+++ project/CHANGELOG.md
@@ -13,3 +13,8 @@ All notable changes to this project will be documented in this file.
 ## [v0.1.0] - 2020-04-24
 ### Changed
 * Integrated chocolatey architecture
+
+## [v0.1.1] - 2020-04-25
+### Added
+* Firefox on Windows
+* Chrome on Windows